CSS clean Adjustment Routeur and Admin Access

+++ AlaskaBillet/controller/ControllerComment.php
@@ -1,4 +1,5 @@
 <?php
+session_start();
 require_once('controller/ControllerBase.php');
 
 // "Colle entre la vue et modele"
@@ -12,22 +13,28 @@ class ControllerComment extends ControllerBase {
   }
 
   public function ShowComments(){
-    $this->_commentManager = new CommentManager();
-    $listeComm = $this->_commentManager->getAllComments();
-    $this->_view = new View('Comment');
-    $this->_view->generate(array('listeComm' => $listeComm));
+    if($this->verifyAdmin()){
+      $this->_commentManager = new CommentManager();
+      $listeComm = $this->_commentManager->getAllComments();
+      $this->_view = new View('Comment');
+      $this->_view->generate(array('listeComm' => $listeComm)); 
+    } else {
+      throw new ErrorMsg('Vous n\'etes pas Admin');
+    }
   }
 
   public function deleteComment($param){
-    if(isset($param['deleteComm'])){
-      $this->_commentManager = new CommentManager();
-      if($this->_commentManager->delete($param['id'])){
-        echo "Mise a jour effectuee";
-      } else {
-        throw new ErrorMsg("Erreur.");
+    if($this->verifyAdmin()){
+      if(isset($param['deleteComm'])){
+        $this->_commentManager = new CommentManager();
+        if($this->_commentManager->delete($param['id'])){
+          $this->Alert('Le commentaire a bien ete supprime', 'success');
+        } else {
+          throw new ErrorMsg("Erreur lors de la suppression du commentaire");
+        }
       }
-    }
     $this->ShowComments();
+    }
   }
